Create createRerunProwJobIcon function in common package
This createRerunProwJobIcon creating a rerun button is moved
from the deck module into the common package.

Setting the color of the text in rerun-content box as gray to
preserve cosistency since it is currently handled on higher levels.
